+++ src/models/event/order/index.ts
@@ -1,17 +1,19 @@
-import { Schema, Types, model } from "mongoose";
+import { Schema, model } from "mongoose";
 import { Ingredient } from "karikarihelper";
 
 // Types
-import { InHouseError, Statics } from "@types";
+import { Statics } from "@types";
 
 // Enums
 import { OrderStatus } from "@enums";
 
-// Models
-import { EventModel, OperatorModel, RealmModel } from "@models";
-
-// Services
-import { StringService } from "@services";
+// Validators
+import {
+    validateOperatorRealm,
+    validateOrderClient,
+    validateOrderEvent,
+    validateOrderOperator,
+} from "./validators";
 
 export enum OrderErrors {
     CLIENT_NAME_GREATER_THAN_MAX_LENGTH = "ERROR_ORDER_CLIENT_NAME_GREATER_THAN_MAX_LENGTH",
@@ -29,48 +31,6 @@ export enum OrderErrors {
     REALM_INVALID = "ERROR_OPERATOR_REALM_INVALID",
     REALM_REQUIRED = "ERROR_OPERATOR_REALM_REQUIRED",
 }
-
-const validateOrderEvent = async (eventId: Types.ObjectId) => {
-    const foundEvent = await EventModel.findById(eventId);
-
-    if (!foundEvent) {
-        throw new InHouseError(OrderErrors.EVENT_INVALID);
-    }
-};
-
-const validateOrderOperator = async (operatorId: Types.ObjectId) => {
-    const foundOperator = await OperatorModel.findById(operatorId);
-
-    if (!foundOperator) {
-        throw new InHouseError(OrderErrors.OPERATOR_INVALID);
-    }
-};
-
-const validateOperatorRealm = async (realmId: Types.ObjectId) => {
-    const foundRealm = await RealmModel.findById(realmId);
-
-    if (!foundRealm) {
-        throw new InHouseError(OrderErrors.REALM_INVALID);
-    }
-};
-
-const validateOrderClient = async (client: string) => {
-    if (
-        StringService.isStringInsideBoundaries(
-            client,
-            Statics.ORDER_CLIENT_NAME_MIN_LENGTH,
-            Statics.ORDER_CLIENT_NAME_MAX_LENGTH
-        ) === false
-    ) {
-        if (client.trim().length < Statics.ORDER_CLIENT_NAME_MIN_LENGTH) {
-            throw new InHouseError(
-                OrderErrors.CLIENT_NAME_LESS_THAN_MIN_LENGTH
-            );
-        }
-
-        throw new InHouseError(OrderErrors.CLIENT_NAME_GREATER_THAN_MAX_LENGTH);
-    }
-};
 
 const OrderSchema = new Schema(
     {
