+++ src/models/index.ts
@@ -1,4 +1,4 @@
-export { default as EventModel, EventErrors } from "./event/event";
+export { default as EventModel, EventErrors } from "./event";
 export { default as MenuModel, MenuErrors } from "./menu";
 export { default as OperatorModel, OperatorErrors } from "./operator";
 export { default as OrderModel, OrderErrors } from "./event/order";
