+++ src/models/product/index.ts
@@ -1,14 +1,11 @@
-import { Schema, Types, model } from "mongoose";
+import { Schema, model } from "mongoose";
 import { Ingredient } from "karikarihelper";
 
 // Types
-import { InHouseError, Statics } from "@types";
+import { Statics } from "@types";
 
-// Models
-import { RealmModel } from "@models";
-
-// Services
-import { DatabaseService, StringService } from "@services";
+// Validators
+import { validateProductName, validateProductRealm } from "./validators";
 
 export enum ProductErrors {
     INVALID = "ERROR_PRODUCT_INVALID",
@@ -21,30 +18,6 @@ export enum ProductErrors {
     REALM_REQUIRED = "ERROR_PRODUCT_REALM_REQUIRED",
     PARENT_INVALID = "ERROR_PRODUCT_PARENT_INVALID",
 }
-
-const validateProductName = async (name: string) => {
-    if (
-        StringService.isStringInsideBoundaries(
-            name,
-            Statics.PRODUCT_NAME_MIN_LENGTH,
-            Statics.PRODUCT_NAME_MAX_LENGTH
-        ) === false
-    ) {
-        if (name.trim().length < Statics.PRODUCT_NAME_MIN_LENGTH) {
-            throw new InHouseError(ProductErrors.NAME_LESS_THAN_MIN_LENGTH);
-        }
-
-        throw new InHouseError(ProductErrors.NAME_GREATER_THAN_MAX_LENGTH);
-    }
-};
-
-const validateProductRealm = async (realmId: Types.ObjectId) => {
-    const foundRealm = await RealmModel.findById(realmId);
-
-    if (!foundRealm) {
-        throw new InHouseError(ProductErrors.REALM_INVALID);
-    }
-};
 
 const ProductSchema = new Schema({
     name: {
